Add item processing stability to not remove excluded items that are required
